#18294 saving progress to the BrowserAPI refactor
